Fix a memory leak in test_rcl_lifecycle. (#1173)
This one came about probably as a result of a bad merge.
But essentially we were forcing the srv_change_state
com_interface to be nullptr, but forgetting to save off
the old pointer early enough.  Thus, we could never restore
the old one before we went to "fini", and the memory would
be leaked.  Fix this by remembering the impl pointer earlier.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>